feat: implement Item management with CRUD operations and update Category management to use Integer IDs

+++ src/main/java/edu/icet/ecom/controller/CategoryController.java
@@ -3,7 +3,6 @@ package edu.icet.ecom.controller;
 import edu.icet.ecom.dto.CategoryDto;
 import edu.icet.ecom.service.CategoryService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.parameters.P;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -26,12 +25,12 @@ public class CategoryController {
     }
 
     @DeleteMapping("/{id}")
-    public boolean deleteById(@PathVariable Long id){
+    public boolean deleteById(@PathVariable Integer id){
         return service.deleteById(id);
     }
 
     @GetMapping("/{id}")
-    public CategoryDto searchById(@PathVariable Long id){
+    public CategoryDto searchById(@PathVariable Integer id){
         return service.searchById(id);
     }
